% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anticholinergic.R
\name{anticholinergic}
\alias{anticholinergic}
\title{Anticholinergic}
\usage{
anticholinergic(df = data.frame())
}
\arguments{
\item{df}{dataframe of size n x p}
}
\value{
an anticho object
}
\description{
Anticholinergic data analysis
}
